Tidy up of cpp examples
Updated rtaudio, removed portaudio (project not updated for a while), all examples in individual folders with individual cmake files, examples fixed and tested
fixes #113, fixes #112, fixes #109, fixes #73

+++ src/maximilian.cpp
@@ -53,9 +53,9 @@ float chandiv= 1;
 
 maxiSettings::maxiSettings() {}
 
-int maxiSettings::sampleRate = 44100;
-int maxiSettings::channels = 2;
-int maxiSettings::bufferSize = 1024;
+size_t maxiSettings::sampleRate = 44100;
+size_t maxiSettings::channels = 2;
+size_t maxiSettings::bufferSize = 1024;
 
 
 //this is a 514-point sinewave table that has many uses.
@@ -692,7 +692,7 @@ bool maxiSample::read()
             }
         }
         amplitudes.resize(shortAmps.size());
-        for(int i=0; i < shortAmps.size(); i++) {
+        for(size_t i=0; i < shortAmps.size(); i++) {
             amplitudes[i] = shortAmps[i] / 32767.0;
         }
         position = amplitudes.size();
@@ -717,7 +717,7 @@ bool maxiSample::save(string filename)
     fstream myFile (filename.c_str(), ios::out | ios::binary);
 
     vector<short> shortAmps(amplitudes.size());
-    for(int i=0; i < shortAmps.size(); i++) {
+    for(size_t i=0; i < shortAmps.size(); i++) {
         shortAmps[i] = static_cast<short>(round(amplitudes[i] * 32767.0));
     }
     // write the wav file per the wav file format
@@ -748,7 +748,7 @@ bool maxiSample::save(string filename)
 //This plays back at the correct speed. Always loops.
 double maxiSample::play() {
     position++;
-    if ((long) position >= amplitudes.size()) position=0;
+    if ((size_t) position >= amplitudes.size()) position=0;
     output = amplitudes[(long)position];
     return output;
 }
@@ -1069,7 +1069,7 @@ void maxiSample::normalise(double maxLevel) {
 
 void maxiSample::autoTrim(float alpha, float threshold, bool trimStart, bool trimEnd) {
 
-    int startMarker=0;
+    size_t startMarker=0;
     if(trimStart) {
         maxiLagExp<double> startLag(alpha, 0);
         while(startMarker < amplitudes.size()) {
@@ -1081,7 +1081,7 @@ void maxiSample::autoTrim(float alpha, float threshold, bool trimStart, bool tri
         }
     }
 
-    int endMarker = amplitudes.size()-1;
+    size_t endMarker = amplitudes.size()-1;
     if(trimEnd) {
         maxiLagExp<float> endLag(alpha, 0);
         while(endMarker > 0) {
@@ -1095,7 +1095,7 @@ void maxiSample::autoTrim(float alpha, float threshold, bool trimStart, bool tri
 
     cout << "Autotrim: start: " << startMarker << ", end: " << endMarker << endl;
 
-    int newLength = endMarker - startMarker;
+    size_t newLength = endMarker - startMarker;
     if (newLength > 0) {
         vector<double> newAmps(newLength);
         for(int i=0; i < newLength; i++) {
